# performance(parser) - Optimize Parser Init