--- a/packages/java-parser/.mocharc.js
+++ b/packages/java-parser/.mocharc.js
@@ -1,0 +1,4 @@
+// Enabled Additional validation flows during Parser / Lexer Initialization.
+process.env["prettier-java-development-mode"] = "enabled";
+
+module.exports = {};