--- a/packages/java-parser/src/lexer.js
+++ b/packages/java-parser/src/lexer.js
@@ -1,9 +1,12 @@
 "use strict";
 const chevrotain = require("chevrotain");
 const { allTokens } = require("./tokens");
+const { getSkipValidations } = require("./utils");
 
 const Lexer = chevrotain.Lexer;
-
-const JavaLexer = new Lexer(allTokens, { ensureOptimizations: true });
+const JavaLexer = new Lexer(allTokens, {
+  ensureOptimizations: true,
+  skipValidations: getSkipValidations()
+});
 
 module.exports = JavaLexer;