Optimize Parser (And Lexer) initialization performance by skipping validation
during productive flows.
The validations are still executed during testing flows using
a mocha configuration file that adds an env variable.

On my Machine the Parser's init time went from 160ms->90ms
And the Lexer's init time went from 30ms->20ms.

Fixes #269