@clement26695 2019-11-23
LGTM.

Just for info: how do you measure the performance ?