@bd82 2019-11-23
> Just for info: how do you measure the performance ?

```javascript
const start = new Date().getTime();
const parser = new Parser();
const end = new Date().getTime();
console.log(end - start);
```